coin mint batching

+++ service/src/main/kotlin/io/provenance/digitalcurrency/consortium/frameworks/AddressDetagQueue.kt
@@ -34,8 +34,8 @@ class AddressDetagQueue(
         start()
     }
 
-    override val numWorkers: Int = coroutineProperties.numWorkers.toInt()
-    override val pollingDelayMillis: Long = coroutineProperties.pollingDelayMs.toLong()
+    override val numWorkers: Int = coroutineProperties.numWorkers
+    override val pollingDelayMillis: Long = coroutineProperties.pollingDelayMs
 
     override suspend fun loadMessages(): List<AddressDetagDirective> =
         transaction {
